added similiarity score for file desc and task desc
